requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1665754496418/work
argon2-cffi @ file:///C:/ci/argon2-cffi_1613038019788/work
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/ci/cffi_1613247308275/work
chardet==4.0.0
click==8.1.3
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1617390489108/work
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
efficientnet-pytorch==0.7.1
entrypoints==0.3
fsspec @ file:///tmp/build/80754af9/fsspec_1617959894824/work
grpcio @ file:///C:/ci/grpcio_1637590993074/work
idna==2.10
imagecodecs @ file:///C:/ci/imagecodecs_1617996781001/work
imageio @ file:///tmp/build/80754af9/imageio_1617700267927/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel==5.5.0
ipython @ file:///C:/ci/ipython_1614616717881/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1610481889018/work
jedi==0.18.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213516947/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
kiwisolver @ file:///C:/ci/kiwisolver_1612282618948/work
locket==0.2.1
lxml==4.7.1
Markdown @ file:///C:/ci/markdown_1614364005059/work
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1601914921407/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx @ file:///tmp/build/80754af9/networkx_1617653298338/work
nltk==3.8
notebook @ file:///C:/ci/notebook_1611348264852/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
olefile==0.46
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas @ file:///C:/ci/pandas_1614711338358/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
partd @ file:///tmp/build/80754af9/partd_1618000087440/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1615224334542/work
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit==3.0.16
protobuf==3.20.1
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.8.1
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets @ file:///C:/ci/pywavelets_1601658407053/work
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole @ file:///tmp/build/80754af9/qtconsole_1612458529756/work
QtPy==1.9.0
regex==2022.10.31
requests==2.25.1
scikit-image==0.18.1
scikit-learn @ file:///C:/ci/scikit-learn_1614446716349/work
scipy @ file:///C:/ci/scipy_1614023124510/work
seaborn @ file:///tmp/build/80754af9/seaborn_1608578541026/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
six @ file:///C:/ci/six_1605205426665/work
tensorboard==1.15.0
tensorboardX==2.5.1
terminado==0.9.2
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
thop==0.0.31.post2005241907
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile @ file:///tmp/build/80754af9/tifffile_1619636090847/work
timm==0.6.7
toolz @ file:///home/linux1/recipes/ci/toolz_1610987900194/work
torch==1.8.0
torchaudio==0.8.0
torchsummary==1.5.1
torchtext==0.9.0
torchvision==0.9.0
tornado @ file:///C:/ci/tornado_1606935947090/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1614982742480/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3==1.26.3
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1651670883478/work
widgetsnbextension==3.5.1
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
